requirements.txt
matplotlib==3.7.1
numpy==1.23.5
pandas==2.0.1
pillow==9.5.0
pip==23.0.1
pyyaml==6.0
scikit-image==0.20.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow-cpu==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
seaborn==0.12.2
transformers
pycryptodome 